requirements.txt
pycryptodome==3.9.9

PIL
pyautogui
opencv-python
keyboard
pystray